+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Bugfixes:
 
   - Use #underscore, not #downcase for help text scope (#140, @atipugin)
   - Radio button labels will now include the disabled class as needed. (#156, @ScottSwezey)
+  - Check box labels will now include the disabled class as needed. (@ScottSwezey)
 
 Features: